requirements.txt
python-dotenv
Flask
Flask-RESTful
pypdf
google-generativeai
pydantic
reportlab 